+++ src/SMAPI/Framework/SCore.cs
@@ -190,24 +190,6 @@ namespace StardewModdingAPI.Framework
             // initialise SMAPI
             try
             {
-#if !SMAPI_3_0_STRICT
-                // hook up events
-                ContentEvents.Init(this.EventManager);
-                ControlEvents.Init(this.EventManager);
-                GameEvents.Init(this.EventManager);
-                GraphicsEvents.Init(this.EventManager);
-                InputEvents.Init(this.EventManager);
-                LocationEvents.Init(this.EventManager);
-                MenuEvents.Init(this.EventManager);
-                MineEvents.Init(this.EventManager);
-                MultiplayerEvents.Init(this.EventManager);
-                PlayerEvents.Init(this.EventManager);
-                SaveEvents.Init(this.EventManager);
-                SpecialisedEvents.Init(this.EventManager);
-                TimeEvents.Init(this.EventManager);
-#endif
-
-                // init JSON parser
                 JsonConverter[] converters = {
                     new ColorConverter(),
                     new PointConverter(),
@@ -262,10 +244,6 @@ namespace StardewModdingAPI.Framework
                 // set window titles
                 this.GameInstance.Window.Title = $"Stardew Valley {Constants.GameVersion} - running SMAPI {Constants.ApiVersion}";
                 Console.Title = $"SMAPI {Constants.ApiVersion} - running Stardew Valley {Constants.GameVersion}";
-#if SMAPI_3_0_STRICT
-                this.GameInstance.Window.Title += " [SMAPI 3.0 strict mode]";
-                Console.Title += " [SMAPI 3.0 strict mode]";
-#endif
             }
             catch (Exception ex)
             {
@@ -375,9 +353,6 @@ namespace StardewModdingAPI.Framework
         private void InitialiseAfterGameStart()
         {
             // add headers
-#if SMAPI_3_0_STRICT
-            this.Monitor.Log($"You're running SMAPI 3.0 strict mode, so most mods won't work correctly. If that wasn't intended, install the normal version of SMAPI from https://smapi.io instead.", LogLevel.Warn);
-#endif
             if (this.Settings.DeveloperMode)
                 this.Monitor.Log($"You have SMAPI for developers, so the console will be much more verbose. You can disable developer mode by installing the non-developer version of SMAPI, or by editing {Constants.ApiConfigPath}.", LogLevel.Info);
             if (!this.Settings.CheckForUpdates)
@@ -439,11 +414,6 @@ namespace StardewModdingAPI.Framework
             int modsLoaded = this.ModRegistry.GetAll().Count();
             this.GameInstance.Window.Title = $"Stardew Valley {Constants.GameVersion} - running SMAPI {Constants.ApiVersion} with {modsLoaded} mods";
             Console.Title = $"SMAPI {Constants.ApiVersion} - running Stardew Valley {Constants.GameVersion} with {modsLoaded} mods";
-#if SMAPI_3_0_STRICT
-            this.GameInstance.Window.Title += " [SMAPI 3.0 strict mode]";
-            Console.Title += " [SMAPI 3.0 strict mode]";
-#endif
-
 
             // start SMAPI console
             new Thread(this.RunConsoleLoop).Start();
@@ -926,14 +896,6 @@ namespace StardewModdingAPI.Framework
                 return false;
             }
 
-#if !SMAPI_3_0_STRICT
-            // add deprecation warning for old version format
-            {
-                if (mod.Manifest?.Version is Toolkit.SemanticVersion version && version.IsLegacyFormat)
-                    SCore.DeprecationManager.Warn(mod.DisplayName, "non-string manifest version", "2.8", DeprecationLevel.PendingRemoval);
-            }
-#endif
-
             // validate dependencies
             // Although dependences are validated before mods are loaded, a dependency may have failed to load.
             if (mod.Manifest.Dependencies?.Any() == true)
@@ -1039,7 +1001,7 @@ namespace StardewModdingAPI.Framework
                             return new ContentPack(packDirPath, packManifest, packContentHelper, this.Toolkit.JsonHelper);
                         }
 
-                        modHelper = new ModHelper(manifest.UniqueID, mod.DirectoryPath, this.Toolkit.JsonHelper, this.GameInstance.Input, events, contentHelper, contentPackHelper, commandHelper, dataHelper, modRegistryHelper, reflectionHelper, multiplayerHelper, translationHelper);
+                        modHelper = new ModHelper(manifest.UniqueID, mod.DirectoryPath, this.GameInstance.Input, events, contentHelper, contentPackHelper, commandHelper, dataHelper, modRegistryHelper, reflectionHelper, multiplayerHelper, translationHelper);
                     }
 
                     // init mod