+++ docs/release-notes.md
@@ -8,6 +8,7 @@ These changes have not been released yet.
 
 * For modders:
   * Added `IContentPack.HasFile` method.
+  * Dropped support for all deprecated APIs.
   * Updated to Json.NET 12.0.1.
 
 ## 2.11.3